+++ duoStrength.js
@@ -54,7 +54,7 @@ const SMALL_BUTTONS_CONTAINER = "_2DR3u";
 const SMALL_BUTTON = "_32WtB _2i-mO _1LZ7U vy3TL _3iIWE _1Mkpg _1Dtxl _1sVAI sweRn _1BWZU _26exN QVrnU";
 const LOCKED_POPOUT = "_1PDfx";
 
-const SKILL_SELECTOR = `[data-test="skill-tree"] [data-test="skill"], [data-test="intro-lesson"]`;
+const SKILL_SELECTOR = `[data-test="skill-tree"] [data-test="skill"], [data-test="intro-lesson"], [data-test="tree-section"] a[href]`;
 const CHECKPOINT_SELECTOR = `[data-test="checkpoint-badge"]`;
 
 const flagYOffsets = {
@@ -1028,11 +1028,22 @@ function addStrengthBars(strengths)
 
 	for (let i=0; i<skillElements.length; i++)
 	{
-		let elementContents = [
-			skillElements[i].querySelector(`[data-test="skill-icon"]`),
-			skillElements[i].firstChild.firstChild.lastChild
-		 ];
-
+		let elementContents;
+		if (skillElements[i].getAttribute("data-test") == "skill")
+		{
+			elementContents = [
+				skillElements[i].querySelector(`[data-test="skill-icon"]`),
+				skillElements[i].firstChild.firstChild.lastChild
+			];
+		}
+		else
+		{
+			// This is bonus skill that has not been purchased so is missing on of the inner containers.
+			elementContents = [
+				skillElements[i].querySelector(`[data-test="skill-icon"]`),
+				skillElements[i].firstChild.lastChild
+			];
+		}
 		/* old way of finding name element before new containers
 
 		// name is a span element, normally it is the last element but if the skill is clicked then a new div is created with the start lesson button etc below the name plate. So need to find the correct span element.